:recycle: (901): move whisper/diarization processors into infrastructure
