+++ mcr-core/mcr_meeting/app/infrastructure/diarization.py
@@ -1,3 +1,4 @@
+import re
 import tempfile
 import time
 from io import BytesIO
@@ -20,9 +21,6 @@ from mcr_meeting.app.schemas.transcription_schema import (
     DiarizationJobStatus,
     DiarizationSegment,
 )
-from mcr_meeting.app.services.speech_to_text.utils import (
-    convert_to_french_speaker,
-)
 from mcr_meeting.app.services.speech_to_text.utils.models import (
     get_diarization_pipeline,
 )
@@ -30,6 +28,11 @@ from mcr_meeting.app.services.speech_to_text.utils.models import (
 api_settings = TranscriptionApiSettings()
 diarization_params = PyannoteDiarizationParameters()
 celery_settings = CelerySettings()
+
+
+def convert_to_french_speaker(speaker_label: str) -> str:
+    """Convert speaker labels to French format"""
+    return re.sub(r"SPEAKER_(\d+)", r"LOCUTEUR_\1", speaker_label)
 
 
 def next_poll_interval(phase_elapsed_s: float, queue_position: int | None) -> float:
