+++ mcr-core/mcr_meeting/app/infrastructure/transcription.py
