+++ mcr-core/mcr_meeting/app/services/speech_to_text/utils/__init__.py
@@ -5,11 +5,9 @@ from .models import (
     get_diarization_pipeline,
     get_transcription_model,
 )
-from .vad import convert_to_french_speaker
 
 __all__ = [
     "split_audio_on_timestamps",
     "get_transcription_model",
     "get_diarization_pipeline",
-    "convert_to_french_speaker",
 ]
