+++ mcr-core/tests/services/speech_to_text_pipeline/test_integration_center_process.py
@@ -92,18 +92,10 @@ def run_the_code_to_test(
         ),
     ],
 )
-@patch(
-    "mcr_meeting.app.services.speech_to_text.diarization_processor.get_feature_flag_client"
-)
-@patch(
-    "mcr_meeting.app.services.speech_to_text.transcription_processor.get_feature_flag_client"
-)
-@patch(
-    "mcr_meeting.app.services.speech_to_text.transcription_processor.get_transcription_model"
-)
-@patch(
-    "mcr_meeting.app.services.speech_to_text.diarization_processor.get_diarization_pipeline"
-)
+@patch("mcr_meeting.app.infrastructure.diarization.get_feature_flag_client")
+@patch("mcr_meeting.app.infrastructure.transcription.get_feature_flag_client")
+@patch("mcr_meeting.app.infrastructure.transcription.get_transcription_model")
+@patch("mcr_meeting.app.infrastructure.diarization.get_diarization_pipeline")
 def test_integration_center_process_normal_flow(
     mock_get_diarization_pipeline: MagicMock,
     mock_get_transcription_model: MagicMock,
@@ -209,12 +201,8 @@ def test_integration_center_process_normal_flow(
         assert transcription_segments[6].speaker == "Intervenant 2"
 
 
-@patch(
-    "mcr_meeting.app.services.speech_to_text.diarization_processor.get_feature_flag_client"
-)
-@patch(
-    "mcr_meeting.app.services.speech_to_text.diarization_processor.get_diarization_pipeline"
-)
+@patch("mcr_meeting.app.infrastructure.diarization.get_feature_flag_client")
+@patch("mcr_meeting.app.infrastructure.diarization.get_diarization_pipeline")
 def test_integration_center_process_empty_diarization(
     mock_get_diarization_pipeline: MagicMock,
     mock_get_feature_flag_client_diarization: MagicMock,
@@ -251,18 +239,10 @@ def test_integration_center_process_empty_diarization(
     assert len(transcription_segments) == 0
 
 
-@patch(
-    "mcr_meeting.app.services.speech_to_text.diarization_processor.get_feature_flag_client"
-)
-@patch(
-    "mcr_meeting.app.services.speech_to_text.transcription_processor.get_feature_flag_client"
-)
-@patch(
-    "mcr_meeting.app.services.speech_to_text.transcription_processor.get_transcription_model"
-)
-@patch(
-    "mcr_meeting.app.services.speech_to_text.diarization_processor.get_diarization_pipeline"
-)
+@patch("mcr_meeting.app.infrastructure.diarization.get_feature_flag_client")
+@patch("mcr_meeting.app.infrastructure.transcription.get_feature_flag_client")
+@patch("mcr_meeting.app.infrastructure.transcription.get_transcription_model")
+@patch("mcr_meeting.app.infrastructure.diarization.get_diarization_pipeline")
 def test_integration_center_process_with_empty_chunks(
     mock_get_diarization_pipeline: MagicMock,
     mock_get_transcription_model: MagicMock,
