+++ mcr-core/tests/services/speech_to_text_pipeline/test_integration_full_process.py
@@ -36,18 +36,10 @@ from mcr_meeting.app.services.speech_to_text.speech_to_text import SpeechToTextP
 )
 @patch("mcr_meeting.app.services.speech_to_text.speech_to_text.AcronymCorrector")
 @patch("mcr_meeting.app.services.speech_to_text.speech_to_text.get_feature_flag_client")
-@patch(
-    "mcr_meeting.app.services.speech_to_text.diarization_processor.get_feature_flag_client"
-)
-@patch(
-    "mcr_meeting.app.services.speech_to_text.transcription_processor.get_feature_flag_client"
-)
-@patch(
-    "mcr_meeting.app.services.speech_to_text.transcription_processor.get_transcription_model"
-)
-@patch(
-    "mcr_meeting.app.services.speech_to_text.diarization_processor.get_diarization_pipeline"
-)
+@patch("mcr_meeting.app.infrastructure.diarization.get_feature_flag_client")
+@patch("mcr_meeting.app.infrastructure.transcription.get_feature_flag_client")
+@patch("mcr_meeting.app.infrastructure.transcription.get_transcription_model")
+@patch("mcr_meeting.app.infrastructure.diarization.get_diarization_pipeline")
 def test_integration_full_process(
     mock_get_diarization_pipeline,
     mock_get_transcription_model,
@@ -199,15 +191,9 @@ def test_integration_full_process(
 
 
 @patch("mcr_meeting.app.services.speech_to_text.speech_to_text.get_feature_flag_client")
-@patch(
-    "mcr_meeting.app.services.speech_to_text.diarization_processor.get_feature_flag_client"
-)
-@patch(
-    "mcr_meeting.app.services.speech_to_text.transcription_processor.get_feature_flag_client"
-)
-@patch(
-    "mcr_meeting.app.services.speech_to_text.diarization_processor.get_diarization_pipeline"
-)
+@patch("mcr_meeting.app.infrastructure.diarization.get_feature_flag_client")
+@patch("mcr_meeting.app.infrastructure.transcription.get_feature_flag_client")
+@patch("mcr_meeting.app.infrastructure.diarization.get_diarization_pipeline")
 def test_integration_full_process_empty_diarization(
     mock_get_diarization_pipeline,
     mock_get_feature_flag_client_transcription,
